due use ord to sort
When a not has two new cards, they usually have the same due
value. The one with the smallest ord is seen first. This was not
reflected in the browser, where instead the cards' order was
determined by sqlite internal decision (e.g. the order of creation of
the card). This change ensure that cards with the same due date are
sorted according to their ord, and so are displayed in the order in
which they will be seen.

Note that this does not deal with the problem of new cards of
different notes having the same due number. But this is also the case
in _fillNew, so I believe it's not something that should be corrected here